Changed all functions to void return type
In case of any error, function calls error() which calls free_data and exits()

Separated files to each to have its own functions

+++ libft/standard/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "../includes/libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int	res;
 	int	minus;
